chore(:necktie:): break user source and tests out into individual files
AFFECTS PACKAGES:
@esri/arcgis-rest-users
