+++ packages/arcgis-rest-users/src/index.ts
@@ -1,4 +1,5 @@
 /* Copyright (c) 2018 Environmental Systems Research Institute, Inc.
  * Apache-2.0 */
 
-export * from "./users";
+export * from "./get";
+export * from "./notification";
